Allow Wizard window to be resizable and have a min/max size
Previously the Wizard window was not resizable and had a hard-coded,
fixed size of 650 x 500. This commit introduces 3 new properties
(ResizeMode, MinSize, MaxSize) that allow a subclass of WizardBase to
make the Wizard window resizable and/or to set the window's minimum and
maximum size. This added flexibility is useful for wizard
implementations that consist of many pages or that contain pages that
require a larger client area.

The default behaviour of Orc.Wizard remains unchanged, i.e. the default
values for the 3 new properties are equal to the values previously
hard-coded into WizardWindow.xaml.