Add optional "Help" button to bottom-left of wizard window
The "Help" button is not visible by default. If it is made visible by
a wizard implementation (property IsHelpVisible), it displays with the
label "Help" and is enabled by default. If necessary the implementation
can disable the button (property CanShowHelp). When the user clicks the
button the event nothing happens by default. It is the responsibility
of the implementation to handle the event, either by reacting to the
event HelpShown or by overriding ShowHelpAsync().

"Creative" wizard implementations can make the button display a
different label than "Help" and do something completely else. The only
thing that is fixed is the button's location in the bottom-left of
the wizard window.